Refactor: add one more layer of validation to the Channel constructor
From RFC:

Channel names are strings (beginning with specified prefix characters).
Apart from the requirement of the first character being a valid channel
type prefix character; the only restriction on a channel name is that it
may not contain any spaces (' ', 0x20), a control G / BELL ('^G', 0x07),
or a comma (',', 0x2C) (which is used as a list item separator by the protocol).